requirements.txt
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.34.0
plotly>=5.20.0
python-dotenv>=1.0.1
requests>=2.31.0
